// src/api-generated/model/createProductDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SaaS API
 * The core API for my multi-tenant SaaS
 * OpenAPI spec version: 1.0
 */
import type { CreateProductDtoType } from './createProductDtoType';
import type { CreateProductDtoSpecifications } from './createProductDtoSpecifications';
import type { ProductAddonDto } from './productAddonDto';
import type { ProductMediaDto } from './productMediaDto';
import type { ProductInstanceDto } from './productInstanceDto';

export interface CreateProductDto {
  type: CreateProductDtoType;
  title: string;
  description?: string;
  base_price: number;
  currency?: string;
  specifications?: CreateProductDtoSpecifications;
  available_addons?: ProductAddonDto[];
  media?: ProductMediaDto[];
  instances?: ProductInstanceDto[];
}